% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromOverlay.R
\name{addPoints}
\alias{addPoints}
\title{Add points}
\usage{
addPoints(poly, steps = 5)
}
\arguments{
\item{poly}{a data frame containing ordered coordinates with polygon vertices}

\item{steps}{numeric, the number of points that should be added between
each point. Default is 10.}
}
\value{
A data frame containing densified coordinates for polygon vertices and
   any other column (assuming original columns contained unique values).
}
\description{
Increases the number of points in a polygon while maintaining the shape
}
\details{
Internal use only.
}
\author{
Giuseppe D'Agostino
}
