% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas.R
\name{prepAtlas}
\alias{prepAtlas}
\title{Prepare a SingleCellExperiment Atlas}
\usage{
prepAtlas(x, res = 300, labels, dimred = NULL, as_map = FALSE)
}
\arguments{
\item{x}{one of the following: a `SingleCellExperiment` class object, a `data.frame` or a `matrix`}

\item{res}{numeric, the resolution of the boundary estimation via `oveRlay::makeOverlay()`. Default is 300}

\item{labels}{character scalar, the name of the column in `colData(x)` with labels, 
if x is a `SingleCellExperiment` class object. Otherwise, a character vector or a factor}

\item{dimred}{character, the name of the reduced dimension in `reducedDim(x)` to be used as basis. 
Only used if x is a `SingleCellExperiment` class object.}

\item{as_map}{logical, should the coordinates be changed to accommodate a geographical projection? Default is FALSE}
}
\value{
a list containing the following slots:
   `atlas` - the atlas coordinates and labels;
   `dr` - a data.frame containing coordinates in 2D embedding of choice
   `coords` - a data.frame containing coordinates for label positioning
   `res` - the resolution used
}
\description{
Prepares a 2D atlas for plotting
}
\author{
Giuseppe D'Agostino
}
