% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromOverlay.R
\name{smoothPolygon}
\alias{smoothPolygon}
\title{Smooth polygons}
\usage{
smoothPolygon(poly, smoothness = 3, min_points = 8, n_dense = 10)
}
\arguments{
\item{poly}{a data frame containing ordered coordinates with polygon vertices}

\item{smoothness}{numeric, the extent of kernel smoothing. Higher means
rounder shapes. Default is 3.}

\item{min_points}{numeric, the minimum number of vertices to smooth.
Default is 8.}

\item{n_dense}{numeric, the number of points to add to the polygon for more
smoothing. Default is 10.}
}
\value{
A data frame containing ordered coordinates for polygon vertices and
   three columns indicating whether they are a hole ("inner") or not ("outer"),
   to which cluster they belong, and a sub-clustering to allow \code{ggplot2}
   to draw them as holes.
}
\description{
Uses kernel smoothing to smooth polygons
}
\details{
This is a refactoring of `smoothr::smooth_ksmooth()` to isolate the
   necessary code and avoid heavy GDAL-based dependencies. The code has been
   simplified assuming `wrap = TRUE` and adding some other bits to handle DF
   with other columns as input.
}
\author{
Matthew Strimas-Mackey, modified by Giuseppe D'Agostino
}
