% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas.R
\name{plotAtlas}
\alias{plotAtlas}
\title{Plot a SingleCellExperiment Atlas}
\usage{
plotAtlas(
  atlas_ret,
  plot_cells = TRUE,
  add_contours = TRUE,
  show_labels = TRUE,
  as_map = FALSE,
  map_proj = "lagrange",
  map_theme = "classic",
  pal = NULL,
  capitalize_labels = FALSE
)
}
\arguments{
\item{atlas_ret}{a list as returned by \code{prepAtlas()}}

\item{plot_cells}{logical, should cells/points be plotted? Default is TRUE}

\item{add_contours}{logical, should density contours be drawn? Default is TRUE}

\item{show_labels}{logical, should labels be plotted? Default is TRUE}

\item{as_map}{logical, should the coordinates be changed to accommodate a geographical projection? Default is FALSE}

\item{map_proj}{character, the map projection. Accepts any one character argument to \code{ggplot2::coord_map()}}

\item{map_theme}{character, one of "classic" (default), "renaissance", "medieval", "modern"}

\item{pal}{character, a vector of colors. Default is NULL and decided by the theme}

\item{capitalize_labels}{logical, should labels be capitalized? Default is FALSE}
}
\value{
a \code{ggplot2} plot object with an atlas-like representation of a 2D point cloud
}
\description{
Plots a 2D atlas from a SingleCellExperiment
}
\author{
Giuseppe D'Agostino
}
