% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas.R
\name{prepAtlas}
\alias{prepAtlas}
\title{Prepare a SingleCellExperiment Atlas}
\usage{
prepAtlas(x, res = 300, labels, dimred = NULL, as_map = FALSE)
}
\arguments{
\item{x}{a \code{SingleCellExperiment} class object, OR a \code{matrix} with 2 columns, OR a \code{data.frame} with 2 columns.}

\item{res}{numeric, the resolution of the boundary estimation via \code{oveRlay::makeOverlay()}. Default is 300}

\item{labels}{character, the name of the column in \code{colData(sce)} with labels
OR a character vector of labels of length \code{nrow(x)} if is a \code{matrix} or a \code{data.frame}}

\item{dimred}{character, the name of the reduced dimension in \code{reducedDim(sce)} 
to be used as basis. Only used if \code{x} is a \code{SingleCellExperiment}.}

\item{as_map}{logical, should the coordinates be changed to accommodate a 
geographical projection? Default is FALSE}
}
\value{
a list containing the following slots:
   `atlas` - the atlas coordinates and labels;
   `dr` - a data.frame containing coordinates in 2D embedding of choice
   `coords` - a data.frame containing coordinates for label positioning
   `res` - the resolution used
}
\description{
Prepares a 2D atlas for plotting
}
\author{
Giuseppe D'Agostino
}
