% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas.R
\name{prepAtlas}
\alias{prepAtlas}
\title{Prepare a SingleCellExperiment Atlas}
\usage{
prepAtlas(sce, res = 300, labels, dimred, as_map = FALSE)
}
\arguments{
\item{sce}{a SingleCellExperiment class object}

\item{res}{numeric, the resolution of the boundary estimation via `oveRlay::makeOverlay()`. Default is 300}

\item{labels}{character, the name of the column in `colData(sce)` with labels}

\item{dimred}{character, the name of the reduced dimension in `reducedDim(sce)` to be used as basis}

\item{as_map}{logical, should the coordinates be changed to accommodate a geographical projection? Default is FALSE}
}
\value{
a list containing the following slots:
   `atlas` - the atlas coordinates and labels;
   `dr` - a data.frame containing coordinates in 2D embedding of choice
   `coords` - a data.frame containing coordinates for label positioning
   `res` - the resolution used
}
\description{
Prepares a 2D atlas for plotting
}
